separate hist functions into two areas
